111: Updating node function tests.  In progress...
