+++ functions/node/jest.config.js
@@ -2,4 +2,14 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
+  collectCoverageFrom: [
+    "**/*.{js,ts}"
+  ],
+  coveragePathIgnorePatterns: [
+    "<rootDir>/node_modules/",
+    "dist/",
+    "coverage/",
+    "lib/adapters/types/",
+    "jest.config.js",
+  ]
 };
