+++ functions/node/tsconfig.json
@@ -6,5 +6,8 @@
     "rootDir": ".",
     "sourceMap": true,
     "strict": false
-  }
+  },
+  "exclude": [
+    "node_modules"
+  ]
 }
